improve tests. Correct schema handling
