+++ csharp/src/Drivers/Apache/Impala/ImpalaStatement.cs
@@ -24,7 +24,6 @@ namespace Apache.Arrow.Adbc.Drivers.Apache.Impala
         internal ImpalaStatement(ImpalaConnection connection)
             : base(connection)
         {
-            ValidateOptions(connection.Properties);
         }
 
         /// <summary>
