+++ csharp/test/Drivers/Apache/Hive2/DateTimeValueTests.cs
@@ -15,34 +15,24 @@
 * limitations under the License.
 */
 
-using System.Collections.Generic;
+using System;
+using System.Globalization;
+using System.Threading.Tasks;
+using Apache.Arrow.Adbc.Drivers.Apache.Spark;
+using Xunit;
 using Xunit.Abstractions;
 
 namespace Apache.Arrow.Adbc.Tests.Drivers.Apache.Hive2
 {
-    public class ClientTests : Common.ClientTests<ApacheTestConfiguration, HiveServer2TestEnvironment>
+    public class DateTimeValueTests : Common.DateTimeValueTests<ApacheTestConfiguration, HiveServer2TestEnvironment>
     {
-        public ClientTests(ITestOutputHelper? outputHelper)
-            : base(outputHelper, new HiveServer2TestEnvironment.Factory())
-        {
-        }
+        public DateTimeValueTests(ITestOutputHelper output)
+            : base(output, new HiveServer2TestEnvironment.Factory())
+        { }
 
-        protected override IReadOnlyList<int> GetUpdateExpectedResults()
+        protected override string GetFormattedTimestampValue(string value)
         {
-            int affectedRows = ValidateAffectedRows ? 1 : -1;
-            return GetUpdateExpectedResults(affectedRows);
-        }
-
-        internal static IReadOnlyList<int> GetUpdateExpectedResults(int affectedRows)
-        {
-            return
-                [
-                    -1, // DROP TABLE
-                    -1, // CREATE TABLE
-                    affectedRows,  // INSERT
-                    affectedRows,  // INSERT
-                    affectedRows,  // INSERT
-                ];
+            return "TO_TIMESTAMP(" + QuoteValue(value) + ")";
         }
     }
 }
